Update patch set 4
Patch Set 4:

I get increased battery drain all the time with it. I think it's interacting with my libmdmcutback.so (which is the lib that actually links against the missing symbol, at least for victara).

It looks like the data connection is going up and down constantly or something according to the logs but I don't see that in the interface.

Patch-set: 4